Issue 355 and Issue 367
Issue 355 - Year deletion is possible now. Does not let the user add more than 3 years and delete year 1.
Issue 367 - This would solve the browser security issue. However, this is not tested as we require a server space to test this.
However, this is pushed to branch - Issue 364 for simplicity purposes.