Issue379-Fixing Tables
In this commit the accordion tables were fixed. The changes included: corn after corn tabs now can be expanded and have the correct data in them.  The enterprise link at the bottom was changed from <a> tag to <button> tag because a plus sign would appear after click to deep in the accordion tables. For more information refer to #Issue379.